refactor: moved borg repo info models and cmd execution into backups.py
